refactor: changing firebase notification service to notification service
